removed index.ts and moved main stuff to src/bugsnag.ts
